Floating point lighting

+++ src/main/java/legend/game/combat/effects/LightningBoltEffectSegment30.java
@@ -14,10 +14,10 @@ public class LightningBoltEffectSegment30 {
   public final USCOLOUR innerColourFadeStep_1c = new USCOLOUR();
   public final USCOLOUR outerColourFadeStep_22 = new USCOLOUR();
   /** ubyte */
-  public int scaleMultiplier_28;
+  public float scaleMultiplier_28;
 
   /** short; Incremented while rendering, but never used for anything. */
-  public int unused_2a;
+  public float unused_2a;
   /** short */
   public int originTranslationMagnitude_2c;
   /** short */
